PyOrchestrator/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
litellm>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pyyaml>=6.0
python-dotenv>=1.0.0
